Rework anyxml encoding
It seems JRE's Transformers are broken in the DOM->StAX direction,
such they do not work with external StAX implementations like
woodstox-core.

As it turns out, this is quite unnecessary complexity because
stax2-api gives us XMLStreamReader wrapping a DOM Element, which
we are already using in the parser side.

This patch removes the the use the transformer in favor of rather
simpler forwarding of XMLStreamReader events to the underlying
XMLStreamWriter.

Also move StAX-based transformer to XmlParserStream, as it is now
the only user of StAX->DOM, updating required feature to StAXSource.

Performs plain old DOM->stream translation, hence it can use any
old transformer.

JIRA: YANGTOOLS-926
Change-Id: I25d0210a6b76f7cbbd5956814773d26776cbace6
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>