Remove checkApplicable() overrides
The only time we are overriding checkApplicable() is in the TOUCH
case, which can more easily be handled through applyTouch().

Refactor AbstractNodeContainerModificationStrategy to include
the utility bits from AutomaticLifecycleMixin and move the override.
This reduces the number of implementations of checkApplicable()
from 7 to 2, while increasing the number of implementations of
checkTouchApplicable() from 3 to 7.

JIRA: YANGTOOLS-943
Change-Id: Iaf198ba852ff88d8e6b570b7c88b01f1064bb3a2
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>
(cherry picked from commit 136e34b09aa90ccbe5d20c0035fb3d573c45daeb)