Add RFC7952 data model extensions
RFC7952 has implications on NormalizedNodeStreamWriter, namely
the ability to emit metadata attached to NormalizedNodes.

This patch introduces the concept of NormalizedMetadata, which
is the metadata counterpart to NormalizedNode. It also defines
NormalizedMetadataStreamWriter as an extension of
NormalizedNodeStreamWriter and implements NormalizedMetadataWriter,
which piggy-backs a NormalizedMetadata structure on top of a
NormalizedNode structure, so that the two are emitted as a single
event stream.

JIRA: YANGTOOLS-961
Change-Id: I8f1a69361d18cf5f9c14185778cca7c2d6ebc4f7
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>