Add NETCONF/RFC7952 compatibility
RFC6241 pre-dates RFC7952 and defines an attribute which should
be modeled as a metadata annotation.

For legacy reasons, we are carrying support for this attribute
in yang-data-api, which does not quite fit the semantic model
and requires special-casing.

Since ietf-netconf.yang does not define the metadata annotation,
which would make RFC7952 work seamlessly, we need to hack this
around a bit and make the correspoding schema definition available.

Change-Id: Ib6c98e9bb8a7c9ae8351b61522c3fbfdd943393c
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>