Deprecate preliminary XPath interfaces
We have our own XPath parser implementation, which allows introspection
into the content and allows further transformations.

The APIs in yang.data.api.schema.xpath assume the XPath parser and
evaluator are kept in the same artifact, which does not work well
with the need to parse XPaths in yang-model-api context and then
evaluating them either in yang-model-api context or in yang-data-api
context.

This patch deprecates those APIs and creates a staging artifact for
APIs which integrate with yang-xpath-api.

yang-data-jaxen is relegated to an experimental feature, pending its
update to the newly-defined APIs.

JIRA: YANGTOOLS-967
Change-Id: I661d50e7024fad3e3bc1f06d8650e506bc061fb7
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>