Define AnyxmlNode and ForeignDataNode
ForeignDataNode is a NormalizedNode anchor pointing towards values
in foreign object models. A prime example is AnyXmlNode, which holds
a value in DOMSource object model.

Since anyxml data can potentially be held in multitude of formats,
define AnyxmlNode which binds ForeignDataNode as a holder of anyxml
value, but does not further specify object model in use.

Finally AnyXmlNode is refactored to be a DOMSource specialization
of AnyxmlNode.

JIRA: YANGTOOLS-975
Change-Id: I1a5476525d711e3e782ed5f8a0691a9fa51f827f
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>