Allow using ImmutableLists when returning binding objects
The general contract of MD-SAL is that we do not leak mutable
collections, as that leads to concurrency issues and provides
an avenue for inconsistencies leaking in.

Unfortunately we have had three offenders, which were using
plain ArrayLists to represent data -- and thus could be used
to subvert our efforts.

This patch adds the ability to use proper ImmutableLists, so that
the data they produce is truly immutable. This also helps footprint
of single-value lists, as those are squashed to a more efficient
implementation.

The default remains the use of mutable lists, but the immutable
versions can be activated through setting system property

org.opendaylight.mdsal.binding.dom.codec.impl.compat-mutable-lists

The property acts as an opt-out, i.e. it is in effect unless it is
specifically set to "false"

JIRA: MDSAL-446
Change-Id: I9d7ebab827314b66539534f8d8200a64ee9b4aff
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>