Add support for Anydata content normalization
In order to transport anydata payloads between parser-based
representations, we need to normalize it and attach a proper
DataSchemaContextTree/DataSchemaContextNode to the structure,
so that the contained data can be properly interpreted.

This patch adds NormalizedAnydata representation, which captures
this information and teaches the XML codec to handle transaformations
to/from the normalized form.

JIRA: YANGTOOLS-995
Change-Id: I3403db3a0250290505c405b5775601d47aadcbbc
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>