Remove AbstractPathArgument.hashCodeGuard
This patch removes the explicit boolean guard around hashCode,
which guarantees disambiguation between 0 and uncomputed.

The single byte expended here has an actual cost of 7 bytes,
as it is forcing all subclasses to have a padding of 7 with
default settings in most environments.

This changes instance sizes as follows (32bit/64bit):

NodeIdentifier                24(3)/32(7) -> 24(4)/24(0)
NodeIdentifierWithPredicates  32(7)/40(7) -> 24(0)/32(0)
NodeWithValue                 32(7)/40(7) -> 24(0)/32(0)

Most notable here is the reduction of
NodeIdentifierWithPredicates, which cannot reasonably be
shared, as it depends on user data.

Change-Id: Ia444cf708f5209cb22f2661e48f2f2ffaf150281
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>