Refactor MountPoint-related definitions
At data-api level we have an opinionated view of how
the underlying mount point looks like and unify the semantics
of a mount point.

Notably we have a MountPointContext, which provides an entrypoint
into the mount point resolution process, which can be dynamically
constructed from inlinee parser data.

This simplifies things by quite a bit, getting rid of unneeded
abstractions in the process.

JIRA: YANGTOOLS-1007
Change-Id: I6419d13b8a84a4cdf5cc7707dba039c33f8d3f39
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>