Proper handling of empty anyData content normalization
Anydata content was always parsed starting from second element,
(to exclude anyData tags), for empty anydata case the closing anyData
tag was passed to XmlParserStream as beginning of a content,
causing ISE as result.

XmlParserStream logic was updated to only parse content beginning with a
starting tag, and omit the other; AnydataNormalizationException is now
thrown for case when there is no anyData content to normalize.

JIRA: YANGTOOLS-1011
Change-Id: I09ae84c44ada198b888d901dff0425b29400fb4f
Signed-off-by: Anna Bencurova <Anna.Bencurova@pantheon.tech>
Signed-off-by: Ruslan Kashapov <ruslan.kashapov@pantheon.tech>