Allow QueuedNotificationManager to work on equality
There are use cases where we would like to use QNM on equality, i.e.
we do not realy have a listener instance, but rather a key to somewhere
else. That other place operates on equality, which makes it incompatible
with the current semantics.

Split the create() method into two, one for identity-based lookups, the
other for equality-based lookups. The power is now in the user's hands
and whatever they do, it really is up to them.

JIRA: YANGTOOLS-1016
Change-Id: I1123d21664a20ff380ccb7db7096801915491f6e
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>