Switch encoding unsigned type normalization to Uint* types
Using Java native types for unsigned values is wasteful especially
in case of uint32 and uint64. Furthermore it leads to uncertainty
between int16/uint8, int32/uint16 and int64/uint32.

This is a partial switch, where we retain java.lang types for
signed entities, but switch the unsigned ones.

JIRA: YANGTOOLS-1018
Change-Id: I0eeb1d343d77d05c4e00cb6919c128042ab31eb1
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>