WIP: Fix Serializer/Deserializer/Codec concepts
Base Codec/Serializer/Deserializer classes need to express the fact
the an implementation can throw an exception. This is important for
proper use of checked exception in implementation domains -- for example
it is natural to throw a XMLStreamException when executing in the
context of StAX streaming.

This prevents user surprises when an XML stream throws an unhandled
IAE exception.

Change-Id: I5d24a36028ae6aef5b244e13c5f9e8b0643f14f4
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>