Update patch set 6
Patch Set 6: Code-Review-1

Will this have a downstream user?
Is the idea to backport this to nitrogen/carbon?

Overall I think we need to fix the model-api/data-api relationship -- getDefault() should really be returning a processed value bound to what yang-data-api uses, not a String. This is important for leafref, as that can cross a lot of models.

Patch-set: 6