Update patch set 1
Patch Set 1: Code-Review-1

Yes, definitely.

model.api needs to reflect the effective RFC7950 world. I think model.api.stmt should be, too.

Also the eval() does not work: for model.api it will always evaluate to 'true', as the if-feature needs to be satisfied for this statement to be visible. For model.api.stmt, I would prefer this to have in a .util package and it needs to get a Collection of features against which it is to be evaluated.

Patch-set: 1
Label: Code-Review=-1