Correct deref() leafref handling
deref() invocation requires the target node to actually be examined,
derefenced and any subsequent path evaluated on top of it.

We correct the implementation to account for correct derefence and
add a UT for a confusing test case. This test case also seeds a new
new yang-model-sa component, which provides methods related to static
analysis of a SchemaContext.

// FIXME: add missing stubs for simple resolution

JIRA: YANGTOOLS-1050
JIRA: YANGTOOLS-1052
Change-Id: I0777da339577f93ecc2786f5beab40b29554fe0e
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>