WIP: Add yang-file-api
In order to deal with YANG module packaging, we need to define some
concepts and interfaces.

Most of these are transplanted from yang.model.repo.api, where act
as companions to higher-level effective model resolution APIs. The
reasoning for their movement is that while they do relate to
yang-model-api, they are not inherently tied to a particular
Java/YANG object model (as defined by yang-model-api).

Change-Id: Ie2d30e2647c02a96dd96ed09c471545e29bee65c
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>