Make BitsTypeDefinition.Bit only a DocumentedNode.WithStatus
The concept that a bit is a SchemaNode directly contradicts RFC7950,
which does not list it as such. Yes, it does have a name that is
an identifier, but it really does not have a namespace.

Fix this issue, which ends up improving our memory footprint by quite
a bit.

JIRA: YANGTOOLS-1072
Change-Id: Id88e14743f9be43f7e4096728bf3715c38470e1a
Signed-off-by: Robert Varga <robert.varga@pantheon.tech>